Add files via upload
This is the updated code for the password strength checker and password generator, that has a GUI. It uses Tkinter to create a GUI that accepts user inputs and uses StringVar() to allow automatic updates from entries into fields. A password scoring system was created and the score is shown only for testing purposes. Maximum password length is used to stop hypothetical DDoS attacks and memory exhaustion since there is no hashing algorithm used that caps a character limit automatically. Regex used to form search patterns that evaluate password strength scores.